[Refactor] presentAlertVC 메소드 UIViewController+로 이동 (#297)
