+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/Base/BaseViewController.swift
@@ -102,14 +102,5 @@ class BaseViewController: UIViewController {
         backgroundView.removeFromSuperview()
         lottieView.removeFromSuperview()
     }
-    
-    func presentAlertVC(title: String, message: String? = nil) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        let alertAction = UIAlertAction(title: StringLiterals.Alert.confirm, style: .cancel) { _ in
-            self.navigationController?.popToRootViewController(animated: false)
-        }
-        alert.addAction(alertAction)
-        self.present(alert, animated: true)
-    }
-    
+
 }
